Allow rrweb breadcrumb customization from hybrid SDKs
