+++ sentry-android-replay/src/main/java/io/sentry/android/replay/DefaultReplayBreadcrumbConverter.kt
@@ -1,0 +1,148 @@
+package io.sentry.android.replay
+
+import io.sentry.Breadcrumb
+import io.sentry.ReplayBreadcrumbConverter
+import io.sentry.SentryLevel
+import io.sentry.SpanDataConvention
+import io.sentry.rrweb.RRWebBreadcrumbEvent
+import io.sentry.rrweb.RRWebEvent
+import io.sentry.rrweb.RRWebSpanEvent
+
+public open class DefaultReplayBreadcrumbConverter : ReplayBreadcrumbConverter {
+    internal companion object {
+        private val snakecasePattern = "_[a-z]".toRegex()
+        private val supportedNetworkData = setOf(
+            "status_code",
+            "method",
+            "response_content_length",
+            "request_content_length",
+            "http.response_content_length",
+            "http.request_content_length"
+        )
+    }
+
+    override fun convert(breadcrumb: Breadcrumb): RRWebEvent? {
+        var rrwebBreadcrumb: RRWebBreadcrumbEvent? = null
+
+        var breadcrumbMessage: String? = null
+        var breadcrumbCategory: String? = null
+        var breadcrumbLevel: SentryLevel? = null
+        val breadcrumbData = mutableMapOf<String, Any?>()
+        when {
+            breadcrumb.category == "http" -> {
+                return if (breadcrumb.isValidForRRWebSpan()) breadcrumb.toRRWebSpanEvent() else null
+            }
+
+            breadcrumb.type == "navigation" &&
+                breadcrumb.category == "app.lifecycle" -> {
+                breadcrumbCategory = "app.${breadcrumb.data["state"]}"
+            }
+
+            breadcrumb.type == "navigation" &&
+                breadcrumb.category == "device.orientation" -> {
+                breadcrumbCategory = breadcrumb.category!!
+                val position = breadcrumb.data["position"]
+                if (position == "landscape" || position == "portrait") {
+                    breadcrumbData["position"] = position
+                } else {
+                    return null
+                }
+            }
+
+            breadcrumb.type == "navigation" -> {
+                breadcrumbCategory = "navigation"
+                breadcrumbData["to"] = when {
+                    breadcrumb.data["state"] == "resumed" -> (breadcrumb.data["screen"] as? String)?.substringAfterLast('.')
+                    "to" in breadcrumb.data -> breadcrumb.data["to"] as? String
+                    else -> null
+                } ?: return null
+            }
+
+            breadcrumb.category == "ui.click" -> {
+                breadcrumbCategory = "ui.tap"
+                breadcrumbMessage = (
+                    breadcrumb.data["view.id"]
+                        ?: breadcrumb.data["view.tag"]
+                        ?: breadcrumb.data["view.class"]
+                    ) as? String ?: return null
+                breadcrumbData.putAll(breadcrumb.data)
+            }
+
+            breadcrumb.type == "system" && breadcrumb.category == "network.event" -> {
+                breadcrumbCategory = "device.connectivity"
+                breadcrumbData["state"] = when {
+                    breadcrumb.data["action"] == "NETWORK_LOST" -> "offline"
+                    "network_type" in breadcrumb.data -> if (!(breadcrumb.data["network_type"] as? String).isNullOrEmpty()) {
+                        breadcrumb.data["network_type"]
+                    } else {
+                        return null
+                    }
+
+                    else -> return null
+                }
+            }
+
+            breadcrumb.data["action"] == "BATTERY_CHANGED" -> {
+                breadcrumbCategory = "device.battery"
+                breadcrumbData.putAll(
+                    breadcrumb.data.filterKeys { it == "level" || it == "charging" }
+                )
+            }
+
+            else -> {
+                breadcrumbCategory = breadcrumb.category
+                breadcrumbMessage = breadcrumb.message
+                breadcrumbLevel = breadcrumb.level
+                breadcrumbData.putAll(breadcrumb.data)
+            }
+        }
+        if (!breadcrumbCategory.isNullOrEmpty()) {
+            rrwebBreadcrumb = RRWebBreadcrumbEvent().apply {
+                timestamp = breadcrumb.timestamp.time
+                breadcrumbTimestamp = breadcrumb.timestamp.time / 1000.0
+                breadcrumbType = "default"
+                category = breadcrumbCategory
+                message = breadcrumbMessage
+                level = breadcrumbLevel
+                data = breadcrumbData
+            }
+        }
+        return rrwebBreadcrumb
+    }
+
+    private fun Breadcrumb.isValidForRRWebSpan(): Boolean {
+        return !(data["url"] as? String).isNullOrEmpty() &&
+            SpanDataConvention.HTTP_START_TIMESTAMP in data &&
+            SpanDataConvention.HTTP_END_TIMESTAMP in data
+    }
+
+    private fun String.snakeToCamelCase(): String {
+        return replace(snakecasePattern) { it.value.last().uppercase() }
+    }
+
+    private fun Breadcrumb.toRRWebSpanEvent(): RRWebSpanEvent {
+        val breadcrumb = this
+        return RRWebSpanEvent().apply {
+            timestamp = breadcrumb.timestamp.time
+            op = "resource.http"
+            description = breadcrumb.data["url"] as String
+            startTimestamp =
+                (breadcrumb.data[SpanDataConvention.HTTP_START_TIMESTAMP] as Long) / 1000.0
+            endTimestamp =
+                (breadcrumb.data[SpanDataConvention.HTTP_END_TIMESTAMP] as Long) / 1000.0
+
+            val breadcrumbData = mutableMapOf<String, Any?>()
+            for ((key, value) in breadcrumb.data) {
+                if (key in supportedNetworkData) {
+                    breadcrumbData[
+                        key
+                            .replace("content_length", "body_size")
+                            .substringAfter(".")
+                            .snakeToCamelCase()
+                    ] = value
+                }
+            }
+            data = breadcrumbData
+        }
+    }
+}
